pyproject.toml
[project]
name = "tesseral-fastapi"
version = "0.0.2"
description = "Tesseral SDK for FastAPI"
readme = "README.md"
authors = [
    { name = "Ulysse Carion", email = "ulysse.carion@tesseral.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "pytest>=8.4.0",
    "starlette>=0.46.1",
    "tesseral-sdk>=0.0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.11.3",
]
